Updates on the reamining files I told about
The portuguese classes have been updated for English:
BaseCaches - CacheBase
BaseUtilizadores - UserBase
CacheEvento - Event

all the remaining already had english names.
The content of which on that wasn't already in the outer folder (except
for the Super-Class cache that I changed some words to english also) has
been changed.

GeocachingPOO (login for the Users) ; Server (total Power and access to
all Data-Bases);

Note: Tradicional  (for Tradicional Cache) all the information here is
already in the SuperClass Cache. Question: We can delete it or we better
ask the tutor?