Make changes to tries for GeocachingPOO and more...
Add country city gender bdate questions when registering for a User,
whiles instead of if's implemented for user tries/tentatives;
New Class for ReportedCacheBase created (Was working on it when
commited.)
New class Treasure with methods and constructors for Cache field.

Now its a list of that objects instead of Strings. in "Cache"

Added points to an activity so that can be updated in the statistics. To
do: Add method statistic to sum all points in all months in all
activities (complex structure reminder)