Change method addActivity both on Main and Geocaching Classes
allowing to receive Coordinates as parameter.  In main, The coordinates
wil be sended as null to the Geocaching ebcause they will be simulated.
In GEocaching it checks if the coordinates are null, if so, it simulates
as previously, if there are coordinates , then it calculares the
distance, sets the kms distance in activity as well, and updates